@@ -1,14 +1,11 @@
 import React from 'react';
-import { Provider } from 'react-redux';
 import { HashRouter } from 'react-router-dom';
 import App from './App';
 
-const Root = ({ store }) => (
-  <Provider store={ store }>
-    <HashRouter>
-      <App />
-    </HashRouter>
-  </Provider>
+const Root = () => (
+  <HashRouter>
+    <App />
+  </HashRouter>
 );
 
 export default Root;
